requirements_backup.txt
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
brotli==1.2.0
build==1.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
dataclasses-json==0.6.7
dill==0.4.0
distro==1.9.0
docling==2.68.0
docling-core==2.59.0
docling-ibm-models==3.10.3
docling-parse==4.7.3
durationpy==0.10
et_xmlfile==2.0.0
Faker==40.1.2
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.59.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
gradio==6.3.0
gradio_client==2.0.3
greenlet==3.3.0
groovy==0.1.2
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-google-genai==4.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
latex2mathml==3.78.1
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.2.6
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.23.2
openai==2.15.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
polyfactory==3.2.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pyclipper==1.4.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pylatexenc==2.10
pyparsing==3.3.1
pypdfium2==4.30.0
PyPika==0.50.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.21
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
rank-bm25==0.2.2
rapidocr==3.5.0
rapidocr-onnxruntime==1.4.4
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safehttpx==0.1.7
safetensors==0.7.0
scipy==1.17.0
semantic-version==2.10.0
semchunk==2.2.2
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
win32_setctime==1.2.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
